feat(struct): interpret dash (`-`) as skip
Tiny addition to avoid errors with libraries that automatically adds tags like json to be skipped with `-`.

In my case, it's an orm generator, to which when I introduced gofakeit and fake tag, it automatically added `fake:"-"` which broke faker.

I managed to manually get fixed changing fake:"-" to fake:"skip".

Amazing library, thanks @brianvoe for all the efforts into it.